+++ rpacore/_sqlite.py
@@ -10,7 +10,7 @@ from rpacore._validation import type_error, value_error
 
 SCHEMA_VERSION_TABLE = "rpacore_schema_versions"
 TRANSACTION_SCHEMA_VERSION = 6
-QUEUE_SCHEMA_VERSION = 3
+QUEUE_SCHEMA_VERSION = 4
 _SUPPORTED_SCHEMA_COMPONENTS = {
     "transactions": (TRANSACTION_SCHEMA_VERSION, "transaction"),
     "queue": (QUEUE_SCHEMA_VERSION, "queue"),